+++ src/main/java/me/ajlane/geo/repast/fire/DefaultFireManager.java
@@ -16,26 +16,25 @@ public class DefaultFireManager implements FireManager {
 
   final static Logger logger = Logger.getLogger(DefaultFireManager.class);
 
-  private final FireSpreader fireSpreader;
+  private final FireSpreader<GridPoint> fireSpreader;
   private final FlammabilityChecker<GridPoint> flammabilityChecker;
+  private final Dimensions gridDims;
   private final Poisson distr;
-  private final Double vegetationMoistureParam;
 
   /**
    * @param fireSpreader Object used to spread fire given an initial ignition
    * @param flammabilityChecker Object that indicates whether or not grid cells are flammable (e.g.
    *        depending on land-cover type)
+   * @param gridDims Grid dimensions
    * @param meanNumFiresPerYear Expected number of fires in the landscape in a year
-   * @param vegetationMoistureParam Dimensionless quantity parameterising the amount of moisture in
-   *        the fuel at the time of the fire
    */
-  public DefaultFireManager(FireSpreader fireSpreader,
-      FlammabilityChecker<GridPoint> flammabilityChecker,
-      Double meanNumFiresPerYear, Double vegetationMoistureParam) {
+  public DefaultFireManager(FireSpreader<GridPoint> fireSpreader,
+      FlammabilityChecker<GridPoint> flammabilityChecker, Dimensions gridDims,
+      Double meanNumFiresPerYear) {
     this.fireSpreader = fireSpreader;
     this.flammabilityChecker = flammabilityChecker;
+    this.gridDims = gridDims;
     this.distr = RandomHelper.createPoisson(meanNumFiresPerYear);
-    this.vegetationMoistureParam = vegetationMoistureParam;
   }
 
   /**
@@ -52,17 +51,16 @@ public class DefaultFireManager implements FireManager {
   @ScheduledMethod(start = 1, interval = 1, priority = 1)
   public void startFires() {
     // List<GridPoint> firesStarted = new ArrayList<GridPoint>();
-    Dimensions gridDims = this.fireSpreader.getLct().getDimensions();
     int numFiresToStart = numFires();
     for (int i = 0; i < numFiresToStart; i++) {
-      GridPoint initialIgnitionPoint = findFlammablePoint(gridDims);
+      GridPoint initialIgnitionPoint = findFlammablePoint(this.gridDims);
       if (initialIgnitionPoint == null) {
         logger.warn("Could not find a flammable cell to initialise fire " + i + "of"
             + numFiresToStart);
         break;
       }
       // firesStarted.add(initialIgnitionPoint);
-      this.fireSpreader.spreadFire(initialIgnitionPoint, this.vegetationMoistureParam);
+      this.fireSpreader.spreadFire(initialIgnitionPoint);
     }
   }