+++ src/test/java/me/ajlane/geo/repast/fire/DefaultFireManagerTest.java
@@ -16,6 +16,11 @@ import repast.simphony.valueLayer.GridValueLayer;
 import repast.simphony.valueLayer.IGridValueLayer;
 import repast.simphony.valueLayer.ValueLayer;
 
+/**
+ * This is mainly an integration test
+ *
+ * @author Andrew Lane
+ */
 public class DefaultFireManagerTest {
 
   final static Logger logger = Logger.getLogger(DefaultFireManagerTest.class);
@@ -28,18 +33,19 @@ public class DefaultFireManagerTest {
 
   private IGridValueLayer lct;
   private IGridValueLayer fireCount;
-  private FireSpreader fireSpreader;
+  private FireSpreader<GridPoint> fireSpreader;
   private FlammabilityChecker<GridPoint> flamChecker;
-  private Double fuelMoistureFactor;
+  private Double vegetationMoistureParam;
 
   @Before
   public void setUp() {
     this.lct = getTestLct();
     this.fireCount = new GridValueLayer("FireCount", 0, true, 5, 5);
-    this.fireSpreader =
-        new FireSpreader(lct, fireCount, srCalc, wrCalc, lcfMap, windDirProbMap, windSpeedProbMap);
+    this.vegetationMoistureParam = 0.5; // lambda in thesis notation
     this.flamChecker = new DefaultFlammabilityChecker(this.lct);
-    this.fuelMoistureFactor = 0.25;
+    this.fireSpreader =
+        new DefaultFireSpreader(lct, fireCount, srCalc, wrCalc, flamChecker, lcfMap,
+            windDirProbMap, windSpeedProbMap, vegetationMoistureParam);
   }
 
   private static IGridValueLayer getTestLct() {
@@ -87,20 +93,20 @@ public class DefaultFireManagerTest {
   public void tearDown() {
     this.fireSpreader = null;
     this.lct = null;
-    this.fuelMoistureFactor = null;
+    this.vegetationMoistureParam = null;
   }
 
   @Test
   public void testInit() {
     double meanNumFiresPerYear = 32.0;
-    new DefaultFireManager(this.fireSpreader, this.flamChecker, meanNumFiresPerYear,
-        this.fuelMoistureFactor);
+    new DefaultFireManager(this.fireSpreader, this.flamChecker,
+        this.lct.getDimensions(), meanNumFiresPerYear);
   }
 
   @Test
   public void testNumFires() {
     DefaultFireManager fireManager = new DefaultFireManager(this.fireSpreader, this.flamChecker,
-        10.1, this.fuelMoistureFactor);
+        this.lct.getDimensions(), 10.1);
     int n = fireManager.numFires();
     logger.debug("Num fires sampled: " + n);
     assertTrue(n > 0);
@@ -108,8 +114,8 @@ public class DefaultFireManagerTest {
 
   @Test
   public void testFiresInitiated() {
-    FireManager fireManager = new DefaultFireManager(this.fireSpreader, this.flamChecker, 5.0,
-        this.fuelMoistureFactor);
+    FireManager fireManager = new DefaultFireManager(this.fireSpreader, this.flamChecker,
+        this.lct.getDimensions(), 5.0);
 
     LctProportionAggregator propAggregator = new LctProportionAggregator(this.lct);
     double initPropBurnt = propAggregator.getLctProportions().get(Lct.Burnt);